Move to a common Hypervisor extension for both VirtualBox and Libvirt.
