--- extensions/libvirtext/libvirtext.proto
@@ -1,25 +1,0 @@
-/* VBox.proto: describes VirtualBox specific state objects
- *
- * Author: J. Lowell Wofford <lowell@lanl.gov>
- *
- * This software is open source software available under the BSD-3 license.
- * Copyright (c) 2021, Triad National Security, LLC
- * See LICENSE file for details.
- */
-
-syntax = "proto3";
-package LibVirtExt;
-option go_package = ".;libvirtext";
-
-import "github.com/gogo/protobuf/gogoproto/gogo.proto";
-option (gogoproto.marshaler_all) = true;
-option (gogoproto.unmarshaler_all) = true;
-option (gogoproto.sizer_all) = true;
-option (gogoproto.goproto_registration) = true;
-option (gogoproto.messagename_all) = true;
-
-message VirtualMachine {
-    string server = 1; // libvirt server name
-    string vm_name = 2; // vm name
-    string uuid = 3; // vm uuid 
-}
