+++ modules/libvirt/README.md
@@ -1,4 +1,3 @@
-# vboxmanage module
+# libvirt module
 
-This module can control "power" state of virtualbox VMs via the vboxmanage-rest-api (https://www.npmjs.com/package/vboxmanage-rest-api).
-
+This module can control "power" state of libvirt VMs via the libvirt RPC interface. Currently only local socket connections are supported.
