--- utils/ansible/roles/kraken-state/templates/LibVirtExt.extension.json.j2
@@ -1,5 +1,0 @@
-{
-{{- '    "server": "' ~ ext.server ~ '",\n' if ext.server is defined else "" -}}
-{{- '    "vmName": "' ~ ext.name ~ '",\n' if ext.name is defined else "" -}}
-    "@type": "type.googleapis.com/LibVirtExt.VirtualMachine"
-}
